@@ -2,7 +2,7 @@ import { Iterator } from './Iterator';
 import { ParallelScanOptions } from './namedParameters';
 import { ParallelScanPaginator } from './ParallelScanPaginator';
 import { ZeroArgumentsConstructor } from '@aws/dynamodb-data-marshaller';
-import DynamoDB = require('aws-sdk/clients/dynamodb');
+import { DynamoDB } from "@aws-sdk/client-dynamodb";
 
 /**
  * Iterates over each item returned by a parallel DynamoDB scan until no more
