@@ -3,8 +3,7 @@ import { DynamoDbResultsPage } from './DynamoDbResultsPage';
 import { mergeConsumedCapacities } from './mergeConsumedCapacities';
 import { ParallelScanInput } from './ParallelScanInput';
 import { ScanPaginator } from './ScanPaginator';
-import { ConsumedCapacity, Key } from 'aws-sdk/clients/dynamodb';
-import DynamoDB = require('aws-sdk/clients/dynamodb');
+import {AttributeValue, ConsumedCapacity, DynamoDB} from '@aws-sdk/client-dynamodb';
 
 /**
  * Pagination state for a scan segment for which the first page has not yet been
@@ -22,7 +21,7 @@ export interface UninitializedScanState {
  */
 export interface InitializedScanState {
     initialized: true;
-    LastEvaluatedKey?: Key;
+    LastEvaluatedKey?: {[key: string]: AttributeValue};
 }
 
 export type ScanState = UninitializedScanState|InitializedScanState;
