@@ -1,7 +1,6 @@
 import { ItemIterator } from './ItemIterator';
 import { QueryPaginator } from './QueryPaginator';
-import { QueryInput } from 'aws-sdk/clients/dynamodb';
-import DynamoDB = require('aws-sdk/clients/dynamodb');
+import {DynamoDB, QueryInput} from '@aws-sdk/client-dynamodb';
 
 export class QueryIterator extends ItemIterator<QueryPaginator> {
     constructor(client: DynamoDB, input: QueryInput, limit?: number) {
