@@ -1,7 +1,6 @@
 import { ItemIterator } from './ItemIterator';
 import { ScanPaginator } from './ScanPaginator';
-import { ScanInput } from 'aws-sdk/clients/dynamodb';
-import DynamoDB = require('aws-sdk/clients/dynamodb');
+import {DynamoDB, ScanInput} from '@aws-sdk/client-dynamodb';
 
 export class ScanIterator extends ItemIterator<ScanPaginator> {
     constructor(client: DynamoDB, input: ScanInput, limit?: number) {
